+++ packages/react/src/meter/root/MeterRoot.tsx
@@ -2,10 +2,21 @@
 import * as React from 'react';
 import PropTypes from 'prop-types';
 import { useComponentRenderer } from '../../utils/useComponentRenderer';
-import { useMeterRoot } from './useMeterRoot';
 import { MeterRootContext } from './MeterRootContext';
+import { mergeProps } from '../../merge-props';
 import { BaseUIComponentProps } from '../../utils/types';
+import { formatNumber } from '../../utils/formatNumber';
+import { useLatestRef } from '../../utils/useLatestRef';
+import { valueToPercent } from '../../utils/valueToPercent';
 import { meterStyleHookMapping } from './styleHooks';
+
+function formatValue(value: number, format?: Intl.NumberFormatOptions): string {
+  if (!format) {
+    return formatNumber(value / 100, [], { style: 'percent' });
+  }
+
+  return formatNumber(value, [], format);
+}
 /**
  * Groups all parts of the meter and provides the value for screen readers.
  * Renders a `<div>` element.
@@ -31,17 +42,40 @@ const MeterRoot = React.forwardRef(function MeterRoot(
     ...otherProps
   } = props;
 
-  const { getRootProps, ...meter } = useMeterRoot({
-    'aria-label': ariaLabel,
-    'aria-labelledby': ariaLabelledby,
-    'aria-valuetext': ariaValuetext,
-    format,
-    getAriaLabel,
-    getAriaValueText,
-    max,
-    min,
-    value,
-  });
+  const formatOptionsRef = useLatestRef(format);
+
+  const percentageValue = valueToPercent(value, min, max);
+
+  const formattedValue = formatValue(value, formatOptionsRef.current);
+
+  const propGetter = React.useCallback(
+    (externalProps = {}) =>
+      mergeProps<'div'>(
+        {
+          'aria-label': getAriaLabel ? getAriaLabel(value) : ariaLabel,
+          'aria-labelledby': ariaLabelledby,
+          'aria-valuemax': max,
+          'aria-valuemin': min,
+          'aria-valuenow': percentageValue / 100,
+          'aria-valuetext': getAriaValueText
+            ? getAriaValueText(value)
+            : (ariaValuetext ?? `${percentageValue}%`),
+          role: 'meter',
+        },
+        externalProps,
+      ),
+    [
+      ariaLabel,
+      ariaLabelledby,
+      ariaValuetext,
+      getAriaLabel,
+      getAriaValueText,
+      max,
+      min,
+      value,
+      percentageValue,
+    ],
+  );
 
   const state: MeterRoot.State = React.useMemo(
     () => ({
@@ -53,14 +87,18 @@ const MeterRoot = React.forwardRef(function MeterRoot(
 
   const contextValue: MeterRootContext = React.useMemo(
     () => ({
-      ...meter,
+      max,
+      min,
+      value,
+      percentageValue,
+      formattedValue,
       state,
     }),
-    [meter, state],
+    [max, min, value, percentageValue, formattedValue, state],
   );
 
   const { renderElement } = useComponentRenderer({
-    propGetter: getRootProps,
+    propGetter,
     render: render ?? 'div',
     state,
     className,
@@ -80,7 +118,50 @@ namespace MeterRoot {
     min: number;
   };
 
-  export interface Props extends useMeterRoot.Parameters, BaseUIComponentProps<'div', State> {}
+  export interface Props extends BaseUIComponentProps<'div', State> {
+    /**
+     * The label for the Indicator component.
+     */
+    'aria-label'?: string;
+    /**
+     * An id or space-separated list of ids of elements that label the Indicator component.
+     */
+    'aria-labelledby'?: string;
+    /**
+     * A string value that provides a human-readable text alternative for the current value of the meter indicator.
+     */
+    'aria-valuetext'?: string;
+    /**
+     * Options to format the value.
+     */
+    format?: Intl.NumberFormatOptions;
+    /**
+     * Accepts a function which returns a string value that provides an accessible name for the Indicator component
+     * @param {number} value The component's value
+     * @returns {string}
+     */
+    getAriaLabel?: (value: number) => string;
+    /**
+     * Accepts a function which returns a string value that provides a human-readable text alternative for the current value of the meter indicator.
+     * @param {number} value The component's value to format
+     * @returns {string}
+     */
+    getAriaValueText?: (value: number) => string;
+    /**
+     * The maximum value
+     * @default 100
+     */
+    max?: number;
+    /**
+     * The minimum value
+     * @default 0
+     */
+    min?: number;
+    /**
+     * The current value.
+     */
+    value: number;
+  }
 }
 
 export { MeterRoot };
